requirements.txt
einops==0.6.1
numpy==1.24.3
opencv_python==4.8.0.76
openpyxl==3.0.9
Pillow==10.0.0
scipy
timm
tqdm==4.61.2